Refactorized Android code
