+++ geteduroam/plugins/wifi-eap-configurator/android/src/main/java/com/emergya/wifieapconfigurator/FactoryAndroid.java
@@ -1,0 +1,15 @@
+package com.emergya.wifieapconfigurator;
+
+import android.os.Build;
+
+public class FactoryAndroid {
+
+	public static Android getAndroid(int api) {
+		if (api <= Build.VERSION_CODES.P) {
+			return new AndroidP();
+		} else {
+			return new AndroidR();
+		}
+	}
+
+}
