+++ geteduroam/src/app/app.component.ts
@@ -69,6 +69,8 @@ export class GeteduroamApp {
       await this.associatedNetwork();
       // Add listeners to app
       await this.addListeners();
+
+      // this.getEduroamServices.setTimeToExpire('1606045500000');
     });
   }
